finished breaking down gui parts into smaller ones
