+++ src/gtnh/gui/lib/text_entry.py
@@ -6,10 +6,10 @@ from gtnh.gui.lib.custom_widget import CustomWidget
 
 
 class TextEntry(Frame, CustomWidget):
-    def __init__(self, master: Any, label_text: str, *args: Any, **kwargs: Any):
+    def __init__(self, master: Any, label_text: str, hide_label: bool = False, *args: Any, **kwargs: Any):
         Frame.__init__(self, master, *args, **kwargs)
         CustomWidget.__init__(self, text=label_text)
-
+        self.hide_label: bool = hide_label
         self.label: Label = Label(self, text=self.label_text)
         self.string_var: StringVar = StringVar(self)
         self.entry: Entry = Entry(self, textvariable=self.string_var)
@@ -21,20 +21,26 @@ class TextEntry(Frame, CustomWidget):
         return self.string_var.get()
 
     def grid_forget(self) -> None:
-        self.label.grid_forget()
+        if not self.hide_label:
+            self.label.grid_forget()
         self.entry.grid_forget()
         super().grid_forget()
 
     def grid(self, *args: Any, **kwargs: Any) -> None:  # type: ignore
-        self.label.grid(row=0, column=0, sticky=Position.LEFT)
+        if not self.hide_label:
+            self.label.grid(row=0, column=0, sticky=Position.LEFT)
         self.entry.grid(row=0, column=1, sticky=Position.HORIZONTAL)
         super().grid(*args, **kwargs)
 
     def configure(self, *args: Any, **kwargs: Any) -> None:  # type: ignore
         if "width" in kwargs:
-            self.label.configure(width=kwargs["width"])
+            if not self.hide_label:
+                self.label.configure(width=kwargs["width"])
             self.entry.configure(width=2 * kwargs["width"])
             del kwargs["width"]
+        if "state" in kwargs:
+            self.entry.configure(state=kwargs["state"])
+            del kwargs["state"]
         super().configure(*args, **kwargs)
 
     def reset(self) -> None:
